requirements.txt
arrow==0.14.7
Click==7.0
dj-database-url==0.4.2
Django==1.11.3
django-crispy-forms==1.6.1
django-registration-redux==1.7
django-tinymce==2.8.0
gunicorn==19.7.1
ics==0.6
psycopg2==2.7.1
python-dateutil==2.8.0
python-dotenv==0.6.4
pytz==2019.3
six==1.12.0
TatSu==4.4.0
whitenoise==3.3.0
